refactor(portfolio-contract): reduce contract, flow type scope
 - keep type-guards focussed on external contract interface
 - minimize coupling
